requirements.txt
discord==1.0.1
discord.py[voice]==2.3.2
python-dotenv==0.15.0
youtube-dl==2021.12.17